# addons/Photo-Vault/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Booster Vault API (Non-AI MVP)
  version: 1.0.0
  description: >
    Booster Vault – Memory/Photo Vault (Non-AI MVP). Encrypted media blobs; server stores metadata + signed URLs only.
servers:
  - url: https://api.boostervault.com
    description: Production
  - url: http://localhost:4000
    description: Local
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable error code.
          example: VALIDATION_ERROR
          examples: [VALIDATION_ERROR, RECOVERY_REQUIRED]
        message:
          type: string
          example: Invalid request
        statusCode:
          type: number
          nullable: true
          description: Optional HTTP status code field (when returned by framework default errors).
        error:
          type: string
          nullable: true
          description: Optional short error label (when returned by framework default errors).
        details:
          type: object
          additionalProperties: true
          nullable: true
        requestId:
          type: string
          nullable: true
          description: Request correlation id (from x-request-id or generated).
...
    Export:
      type: object
      required: [id, userId, scopeType, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        scopeType: { type: string, enum: [VAULT, ALBUM, DATE_RANGE] }
        scopeAlbumId: { type: string, format: uuid, nullable: true }
        scopeFrom: { type: string, format: date-time, nullable: true }
        scopeTo: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [QUEUED, RUNNING, READY, FAILED, EXPIRED] }
        errorMessage: { type: string, nullable: true }
        outputObjectKey: { type: string, nullable: true }
        outputByteSize: { type: integer, format: int64, nullable: true }
        readyAt: { type: string, format: date-time, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PaginatedExportsResponse:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Export" }
        nextCursor: { type: string, nullable: true }
        hasMore: { type: boolean }
    SignedUrl:
      type: object
      required: [url, expiresAt, method, headers]
      properties:
        url: { type: string }
        expiresAt: { type: string, format: date-time }
        method: { type: string, enum: [PUT, GET] }
        headers:
          type: object
          additionalProperties: true
...
    CardContactGate:
      type: string
      enum: [OPEN, REQUEST_REQUIRED, HIDDEN]
    CardAttachmentKind:
      type: string
      enum: [ALBUM, MEDIA, LIFE_DOC]
    CardAttachmentResolvedLink:
      type: object
      required: [kind, shareId, shareLink, expiresAt]
      properties:
        kind:
          type: string
          enum: [SHARED_ALBUM]
        shareId: { type: string, format: uuid }
        shareLink: { type: string }
        expiresAt: { type: string, format: date-time }
    CardContactRequestStatus:
      type: string
      enum: [PENDING, APPROVED, DENIED, EXPIRED, REVOKED]
    CardModePublic:
      type: object
      required:
        - modeId
        - cardPublicId
        - slug
        - name
        - contactGate
        - indexingEnabled
        - createdAt
        - updatedAt
      properties:
        modeId: { type: string, format: uuid }
        cardPublicId: { type: string }
        slug: { type: string }
        name: { type: string }
        headline: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        contactGate: { $ref: "#/components/schemas/CardContactGate" }
        indexingEnabled: { type: boolean }
        themeKey: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CardAttachment:
      type: object
      required: [id, kind, refId, sortOrder]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/CardAttachmentKind" }
        refId: { type: string }
        label: { type: string, nullable: true }
        sortOrder: { type: integer }
        expiresAt: { type: string, format: date-time, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }
        resolvedLink:
          nullable: true
          allOf:
            - { $ref: "#/components/schemas/CardAttachmentResolvedLink" }
...
    GetPublicCardModeResponse:
      type: object
      required: [mode, attachments]
      properties:
        mode: { $ref: "#/components/schemas/CardModePublic" }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/CardAttachment" }
    CreateCardContactRequestDto:
      type: object
      required: [requesterName, requesterEmail]
      properties:
        requesterName: { type: string, minLength: 2, maxLength: 80 }
        requesterEmail: { type: string, format: email }
        requesterPhone: { type: string, maxLength: 30, nullable: true }
        message: { type: string, maxLength: 1000, nullable: true }
        captchaToken:
          type: string
          maxLength: 5000
          nullable: true
          description: Turnstile token (required when CAPTCHA is enabled)
    CreateCardContactRequestResponse:
      type: object
      required: [requestId, status, createdAt]
      properties:
        requestId: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/CardContactRequestStatus" }
        createdAt: { type: string, format: date-time }
...
    CardContactRevealResponse:
      type: object
      properties:
        displayName: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
...
    CreateExportRequest:
      type: object
      required: [scopeType]
      properties:
        scopeType: { type: string, enum: [VAULT, ALBUM, DATE_RANGE] }
        scopeAlbumId: { type: string, format: uuid, nullable: true }
        scopeFrom: { type: string, format: date-time, nullable: true }
        scopeTo: { type: string, format: date-time, nullable: true }
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
paths:
...
  /v1/continuity/releases/execute:
    post:
      tags: [Continuity]
      summary: Execute a continuity release
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [packId]
              properties:
                packId: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/continuity/releases/{id}/revoke:
    post:
      tags: [Continuity]
      summary: Revoke a continuity release
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/exports:
    get:
      tags: [Exports]
      summary: List export jobs
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - in: query
          name: cursor
          schema: { type: string, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [exports]
                properties:
                  exports: { $ref: "#/components/schemas/PaginatedExportsResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }

    post:
      tags: [Exports]
      summary: Create export job
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          description: >
            Unique key for ensuring idempotency of this request. If provided and a previous request with the same key succeeded,
            the same response will be returned. If the request body differs, 409 Conflict will be returned.
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateExportRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [export]
                properties:
                  export: { $ref: "#/components/schemas/Export" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/exports/{exportId}:
    get:
      tags: [Exports]
      summary: Get export job
      parameters:
        - in: path
          name: exportId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [export]
                properties:
                  export: { $ref: "#/components/schemas/Export" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

    delete:
      tags: [Exports]
      summary: Delete export job
      parameters:
        - in: path
          name: exportId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: No Content }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/exports/{exportId}/download-url:
    post:
      tags: [Exports]
      summary: Get signed download URL for export ZIP
      parameters:
        - in: path
          name: exportId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [downloadUrl]
                properties:
                  downloadUrl: { $ref: "#/components/schemas/SignedUrl" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # =========================
  # Storage
  # =========================
...
  /v1/card/public/{publicId}/modes/{modeSlug}:
    get:
      tags: [Card]
      summary: Get public card mode view
      operationId: getPublicCardMode
      security: []
      parameters:
        - name: publicId
          in: path
          required: true
          schema: { type: string }
        - name: modeSlug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Public card mode view
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GetPublicCardModeResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/card/public/{publicId}/modes/{modeSlug}/contact-requests:
    post:
      tags: [Card]
      summary: Create public card contact request
      operationId: createPublicCardContactRequest
      security: []
      parameters:
        - name: publicId
          in: path
          required: true
          schema: { type: string }
        - name: modeSlug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCardContactRequestDto" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [request]
                properties:
                  request: { $ref: "#/components/schemas/CreateCardContactRequestResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/card/public/{publicId}/modes/{modeSlug}/contact:
    get:
      tags: [Card]
      summary: Reveal contact (grant-token protected)
      operationId: getPublicCardModeContact
      security: []
      parameters:
        - name: publicId
          in: path
          required: true
          schema: { type: string }
        - name: modeSlug
          in: path
          required: true
          schema: { type: string }
        - name: X-Card-Token
          in: header
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [contact]
                properties:
                  contact: { $ref: "#/components/schemas/CardContactRevealResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...